Configure bulk Expression Atlas suggesters non-destructively
The scripts deleted the search component that stored the suggesters, causing any suggester from Single Cell Expression Atlas to be removed in the process. With this revision, while the suggester isn’t entirely deleted, its configuration is completely cleared first to avoid any parameters to be kept between collection versions. As long as the name `proppertySuggester` is maintained it will work fine.